lang: Move rain_to_string into rain.c
