+++ lib/_pkg.rn
@@ -3,8 +3,8 @@ link "util.c"
 
 import env
 import except
-import string
 
 let print = foreign "rain_ext_print"(val)
 let exit = foreign "rain_exit"(val)
 let panic = foreign "rain_ext_throw"(val)
+let to_str = foreign "rain_to_string"(val)
